Refactor ping controller Update openapi description
