+++ src/controllers/__tests__/statusServicesController.test.ts
@@ -1,5 +1,5 @@
 import { Request } from "express";
-import { getStatusServices } from "../statusServicesController";
+import { getPing, getStatusServices } from "../statusServicesController";
 import mockRes from "../../__mocks__/response";
 import { mockServicesStatus } from "../__mocks__/statusServices";
 
@@ -16,5 +16,18 @@ describe("getStatusServices", () => {
       });
       result.apply(res);
       expect(res.json).toHaveBeenCalledWith(mockServicesStatus);
+  });
+});
+
+describe("getPing", () => {
+  it("should return a success no content response", async () => {
+    const mockRequest = {} as Request;
+    const result = await getPing(mockRequest);
+    expect(result).toEqual({
+      apply: expect.any(Function),
+      kind: "IResponseSuccessNoContent",
+    });
+    result.apply(res);
+    expect(res.status).toHaveBeenCalledWith(204);
   });
 });
