+++ src/controllers/statusServicesController.ts
@@ -1,7 +1,9 @@
 import {
   IResponseErrorInternal,
   IResponseSuccessJson,
-  ResponseSuccessJson
+  IResponseSuccessNoContent,
+  ResponseSuccessJson,
+  ResponseSuccessNoContent
 } from "@pagopa/ts-commons/lib/responses";
 import { Request } from "express";
 
@@ -22,3 +24,13 @@ export const getStatusServices = async (
     ServicesStatus.decode(mockServicesStatus),
     (servicesStatus) => ResponseSuccessJson(servicesStatus)
   );
+
+/**
+ * Returns success no content.
+ *
+ * @param req - Express request object
+ */
+export const getPing = async (
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  _req: Request
+): Promise<IResponseSuccessNoContent> => ResponseSuccessNoContent();
